refactor: move proof classes from contracts to circuits
