+++ packages/contracts/test/userStateTransition.test.ts
@@ -16,10 +16,12 @@ import {
     NUM_EPOCH_KEY_NONCE_PER_EPOCH,
     Circuit,
     defaultEpochTreeLeaf,
+    UserStateTransitionProof,
+    SignupProof,
 } from '@unirep/circuits'
 import { defaultProver } from '@unirep/circuits/provers/defaultProver'
 
-import { EPOCH_LENGTH, UserStateTransitionProof, SignupProof } from '../src'
+import { EPOCH_LENGTH } from '../src'
 import { deployUnirep } from '../deploy'
 
 const signupUser = async (id, unirepContract, attesterId, account) => {
